organize files and logic
