+++ api/shared/azure-cosmosdb-mongodb.ts
